% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc.R
\name{cloc_stats}
\alias{cloc_stats}
\title{Get lines-of-code ('cloc') statistics for a package}
\usage{
cloc_stats(path)
}
\arguments{
\item{path}{Path to locally extracted tarball of package, obtained from
\link{extract_tarball}.}
}
\value{
A 'data.frame' with statistics from the four main directories (where
present) of '/R', '/src', '/inst/include', and 'vignettes'
}
\description{
Get lines-of-code ('cloc') statistics for a package
}
\examples{
\dontrun{
tarball <- "magrittr_2.0.1.tar.gz"
u <- paste0 ("https://cran.r-project.org/src/contrib/",
             tarball)
dest <- file.path (tempdir (), tarball)
download.file (u, dest)
path <- extract_tarball (dest)
cloc_stats (path)
}
}
