% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctags-install.R
\name{ctags_install}
\alias{ctags_install}
\title{Install 'ctags' from a clone of the 'git' repository}
\usage{
ctags_install(bin_dir = NULL)
}
\arguments{
\item{bin_dir}{Prefix to pass to the \code{autoconf} configure command
defining location to install the binary, with default of \verb{/usr/local}.}
}
\description{
'ctags' is installed with this package on both Windows and macOS systems; this
is an additional function to install from source on Unix systems. The
installation requires 'sudo' privileges, and will fail otherwise.
}
