% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctags-test.R
\name{ctags_test}
\alias{ctags_test}
\title{test a 'ctags' installation using the example from
https://github.com/universal-ctags/ctags/blob/master/man/ctags-lang-r.7.rst.in}
\usage{
ctags_test()
}
\description{
test a 'ctags' installation using the example from
https://github.com/universal-ctags/ctags/blob/master/man/ctags-lang-r.7.rst.in
}
