% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats-data-download.R
\name{dl_pkgstats_data}
\alias{dl_pkgstats_data}
\title{Download latest version of 'pkgstats' data}
\usage{
dl_pkgstats_data(current = TRUE, path = tempdir(), quiet = FALSE)
}
\arguments{
\item{current}{If 'FALSE', download data for all CRAN packages ever released,
otherwise (default) download data only for current CRAN packages.}

\item{path}{Local path to download file.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
(Invisibly) A \code{data.frame} of \code{pkgstats} results, one row for each
package.
}
\description{
Download latest version of 'pkgstats' data
}
\seealso{
Other archive: 
\code{\link{pkgstats_fns_from_archive}()},
\code{\link{pkgstats_from_archive}()}
}
\concept{archive}
