% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tarball.R
\name{extract_tarball}
\alias{extract_tarball}
\title{Extract tarball of a package into temp directory and return path to extracted
package}
\usage{
extract_tarball(tarball)
}
\arguments{
\item{tarball}{Full path to local tarball of an R package}
}
\value{
Path to extracted version of package
}
\description{
Extract tarball of a package into temp directory and return path to extracted
package
}
