% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats-package.R
\docType{package}
\name{pkgstats-package}
\alias{pkgstats}
\alias{pkgstats-package}
\title{pkgstats: Metrics of R Packages}
\description{
Metrics of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/pkgstats}
  \item Report bugs at \url{https://github.com/mpadge/pkgstats/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
