% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats.R
\name{pkgstats}
\alias{pkgstats}
\title{Collates statistics from one local tarball}
\usage{
pkgstats(tarball)
}
\arguments{
\item{tarball}{Full path to local tarball of an R package}
}
\value{
List of statistics
}
\description{
Collates statistics from one local tarball
}
\examples{
\dontrun{
tarball <- "magrittr_2.0.1.tar.gz"
u <- paste0 ("https://cran.r-project.org/src/contrib/",
             tarball)
f <- file.path (tempdir (), tarball)
download.file (u, f)
pkgstats (f)
}
}
