% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-data-current.R
\name{pkgstats_cran_current_from_full}
\alias{pkgstats_cran_current_from_full}
\title{Reduce \code{data.frame} of full CRAN archive data to current packages only.}
\usage{
pkgstats_cran_current_from_full(prev_results, results_file = NULL)
}
\arguments{
\item{prev_results}{Result of previous call to this function, if available.
Submitting previous results will ensure that only newer packages not present
in previous result will be analysed, with new results simply appended to
previous results. This parameter can also specify a file to be read with
\code{readRDS()}.}

\item{results_file}{Can be used to specify the name or full path of a \code{.Rds}
file to which results should be saved once they have been generated. The
'.Rds' extension will be automatically appended, and any other extensions
will be ignored.}
}
\value{
A \code{data.frame} object with one row for each package containing
summary statistics generated from the \link{pkgstats_summary} function.
}
\description{
Reduce \code{data.frame} of full CRAN archive data to current packages only.
}
\seealso{
Other archive: 
\code{\link{dl_pkgstats_data}()},
\code{\link{pkgstats_fns_from_archive}()},
\code{\link{pkgstats_fns_update}()},
\code{\link{pkgstats_from_archive}()},
\code{\link{pkgstats_update}()}
}
\concept{archive}
