% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-data-update.R
\name{pkgstats_update}
\alias{pkgstats_update}
\title{Update local \code{data.frame} of \code{pkgstats} data with data from all packages
uploaded to CRAN since creation of local data.}
\usage{
pkgstats_update(
  prev_results = NULL,
  results_file = NULL,
  chunk_size = 1000L,
  num_cores = 1L,
  save_full = FALSE,
  save_ex_calls = FALSE,
  results_path = tempdir()
)
}
\arguments{
\item{prev_results}{Result of previous call either to this function, or to
\link{pkgstats_from_archive}, as a \code{data.frame}  of statistics, with one row
for each package.}

\item{results_file}{Can be used to specify the name or full path of a \code{.Rds}
file to which results should be saved once they have been generated. The
'.Rds' extension will be automatically appended, and any other extensions
will be ignored.}

\item{chunk_size}{Divide large archive trawl into chunks of this size, and
save intermediate results to local files. These intermediate files can be
combined to generate a single \code{prev_results} file, to enable jobs to be
stopped and re-started without having to recalculate all results. These files
will be named \code{pkgstats-results-N.Rds}, where "N" incrementally numbers each
file.}

\item{num_cores}{Number of machine cores to use in parallel, defaulting to
single-core processing.}

\item{save_full}{If \code{TRUE}, full \link{pkgstats} results are saved for each
package to files in \code{results_path}.}

\item{save_ex_calls}{If \code{TRUE}, the results of the \code{external_calls} component
are saved for each package to files in \code{results_path} (only if \code{save_full = FALSE}).}

\item{results_path}{Path to save intermediate files generated by the
\code{chunk_size} parameter described above.}
}
\description{
Note that this function only updates data from current CRAN packages, and
not from archived packages. Any packages which have been updated multiple
times since generation of previous results will only be updated with the
most recent data, and not with data from any intermediate updates.
}
\examples{
# Create fake archive directory with single tarball:
f <- system.file ("extdata", "pkgstats_9.9.tar.gz", package = "pkgstats")
tarball <- basename (f)

archive_path <- file.path (tempdir (), "archive")
if (!dir.exists (archive_path)) {
    dir.create (archive_path)
}
path <- file.path (archive_path, tarball)
file.copy (f, path)
tarball_path <- file.path (archive_path, "tarballs")
dir.create (tarball_path, recursive = TRUE)
file.copy (path, file.path (tarball_path, tarball))
\dontrun{
out <- pkgstats_from_archive (tarball_path)
}
}
\seealso{
Other archive: 
\code{\link{dl_pkgstats_data}()},
\code{\link{pkgstats_fns_from_archive}()},
\code{\link{pkgstats_from_archive}()}
}
\concept{archive}
