% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-stats.R
\name{rd_stats}
\alias{rd_stats}
\title{Stats from '.Rd' files}
\usage{
rd_stats(path)
}
\arguments{
\item{path}{Directory to package being analysed}
}
\description{
Stats from '.Rd' files
}
\examples{
\dontrun{
tarball <- "magrittr_2.0.1.tar.gz"
u <- paste0 (
    "https://cran.r-project.org/src/contrib/",
    tarball
)
f <- file.path (tempdir (), tarball)
download.file (u, f)
# have to extract tarball to call function on source code:
path <- extract_tarball (f)
rd_stats (path)
}
}
\seealso{
Other stats: 
\code{\link{desc_stats}()},
\code{\link{loc_stats}()},
\code{\link{pkgstats_summary}()},
\code{\link{pkgstats}()}
}
\concept{stats}
