% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc.R
\name{whitespace_stats}
\alias{whitespace_stats}
\title{Statistics on white space in code}
\usage{
whitespace_stats(path)
}
\arguments{
\item{path}{Directory to package being analysed}
}
\value{
A list of statistics for each of three directories, 'R', 'src', and
'inst/include', each one having 5 statistics of total numbers of lines,
numbers of empty lines, total numbers of white spaces, total numbers of
characters, and indentation used in files in that directory.
}
\description{
Statistics on white space in code
}
\note{
NA values are returned for directories which do not exist.
}
