% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctags-test.R
\name{ctags_test}
\alias{ctags_test}
\title{test a 'ctags' installation}
\usage{
ctags_test(quiet = TRUE)
}
\arguments{
\item{quiet}{If \code{TRUE}, display on screen whether or not 'ctags' is correctly
installed.}
}
\value{
'TRUE' or 'FALSE' respectively indicating whether or not 'ctags' is
correctly installed.
}
\description{
This uses the example from
\url{https://github.com/universal-ctags/ctags/blob/master/man/ctags-lang-r.7.rst.in}
and also checks the GNU global installation.
}
\examples{
\dontrun{
ctags_test ()
}
}
\seealso{
Other tags: 
\code{\link{ctags_install}()},
\code{\link{tags_data}()}
}
\concept{tags}
