% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc-stats.R
\name{desc_stats}
\alias{desc_stats}
\title{Statistics from DESCRIPTION files}
\usage{
desc_stats(path)
}
\arguments{
\item{path}{Directory to package being analysed}
}
\description{
Statistics from DESCRIPTION files
}
