% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc-stats.R
\name{desc_stats}
\alias{desc_stats}
\title{Statistics from DESCRIPTION files}
\usage{
desc_stats(path)
}
\arguments{
\item{path}{Directory to source code of package being analysed}
}
\value{
A \code{data.frame} with one row and 16 columns extracting various
information from the 'DESCRIPTION' file, include websites, tallies of
different kinds of authors and contributors, and package dependencies.
}
\description{
Statistics from DESCRIPTION files
}
\examples{
f <- system.file ("extdata", "pkgstats_9.9.tar.gz", package = "pkgstats")
# have to extract tarball to call function on source code:
path <- extract_tarball (f)
desc_stats (path)
}
\seealso{
Other stats: 
\code{\link{loc_stats}()},
\code{\link{pkgstats_summary}()},
\code{\link{pkgstats}()},
\code{\link{rd_stats}()}
}
\concept{stats}
