% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats-package.R
\docType{package}
\name{pkgstats-package}
\alias{pkgstats-package}
\alias{_PACKAGE}
\title{pkgstats: Metrics of R Packages}
\description{
Metrics of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgstats/}
  \item \url{https://github.com/ropensci-review-tools/pkgstats}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgstats/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
