% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats-package.R
\docType{package}
\name{pkgstats-package}
\alias{pkgstats-package}
\title{pkgstats: Metrics of R Packages}
\description{
Static code analyses for R packages using the external code-tagging libraries 'ctags' and 'gtags'. Static analyses enable packages to be analysed very quickly, generally a couple of seconds at most. The package also provides access to a database generating by applying the main function to the full 'CRAN' archive, enabling the statistical properties of any package to be compared with all other 'CRAN' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgstats/}
  \item \url{https://github.com/ropensci-review-tools/pkgstats}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgstats/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Other contributors:
\itemize{
  \item Michael Sumner (\href{https://orcid.org/0000-0002-2471-7511}{ORCID}) [contributor]
  \item Jeffrey Hollister (\href{https://orcid.org/0000-0002-9254-9740}{ORCID}) [contributor]
  \item Egor Kotov (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [contributor]
}

}
\keyword{internal}
