% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats.R
\name{pkgstats}
\alias{pkgstats}
\title{Collates statistics from one local tarball}
\usage{
pkgstats(tarball)
}
\arguments{
\item{tarball}{Full path to local tarball of an R package}
}
\value{
List of statistics
}
\description{
Collates statistics from one local tarball
}
