% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats.R
\name{pkgstats}
\alias{pkgstats}
\title{Collates statistics from one local tarball}
\usage{
pkgstats(path = ".")
}
\arguments{
\item{path}{Either a path to a local source repository, or a local '.tar.gz'
file containing code for an R package.}
}
\value{
List of statistics
}
\description{
Collates statistics from one local tarball
}
\examples{
\dontrun{
tarball <- "magrittr_2.0.1.tar.gz"
u <- paste0 (
    "https://cran.r-project.org/src/contrib/",
    tarball
)
f <- file.path (tempdir (), tarball)
download.file (u, f)
pkgstats (f)
}
}
\seealso{
Other stats: 
\code{\link{desc_stats}()},
\code{\link{loc_stats}()},
\code{\link{pkgstats_summary}()},
\code{\link{rd_stats}()}
}
\concept{stats}
