% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-trawl.R
\name{pkgstats_from_archive}
\alias{pkgstats_from_archive}
\title{Trawl a local CRAN archive and extract statistics from all packages}
\usage{
pkgstats_from_archive(
  path,
  archive = TRUE,
  prev_results = NULL,
  results_file = NULL
)
}
\arguments{
\item{path}{Path to local CRAN archive}

\item{archive}{If \code{TRUE}, extract statistics for all packages in the
\verb{/Archive} sub-directory, otherwise only statistics for main \code{tarballs}
directory (that is, current packages only).}

\item{prev_results}{Result of previous call to this function, if available.
Submitting previous results will ensure that only newer packages not present
in previous result will be analysed, with new results simply appended to
previous results.}

\item{results_file}{Can be used to specify the name or full path of a \code{.Rds}
file to which results should be saved once they have been generated. The
'.Rds' extension will be automatically appended, and any other extensions
will be ignored.}
}
\value{
A \code{data.frame} object with one row for each package containing
summary statistics generated from the \link{pkgstats_summary} function.
}
\description{
Trawl a local CRAN archive and extract statistics from all packages
}
