% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-trawl.R
\name{pkgstats_from_archive}
\alias{pkgstats_from_archive}
\title{Trawl a local CRAN archive and extract statistics from all packages}
\usage{
pkgstats_from_archive(path)
}
\arguments{
\item{path}{Path to local CRAN archive}
}
\description{
Trawl a local CRAN archive and extract statistics from all packages
}
