% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-data-update.R
\name{pkgstats_update}
\alias{pkgstats_update}
\title{Update pkgstats` data on GitHub release}
\usage{
pkgstats_update(upload = TRUE)
}
\arguments{
\item{upload}{If \code{TRUE}, upload updated results to GitHub release.}
}
\value{
Local path to directory containing updated results.
}
\description{
Update pkgstats` data on GitHub release
}
\seealso{
Other archive: 
\code{\link{dl_pkgstats_data}()},
\code{\link{pkgstats_cran_current_from_full}()},
\code{\link{pkgstats_fns_from_archive}()},
\code{\link{pkgstats_fns_update}()},
\code{\link{pkgstats_from_archive}()}
}
\concept{archive}
