% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network}
\alias{plot_network}
\title{Plot interactive \pkg{visNetwork} visualisation of object-relationship
network of package.}
\usage{
plot_network(s)
}
\arguments{
\item{s}{Package statistics obtained from \link{pkgstats} function.}
}
\value{
(Invisibly) A \pkg{visNetwork} representation of the package network.
}
\description{
Plot interactive \pkg{visNetwork} visualisation of object-relationship
network of package.
}
