% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-stats.R
\name{rd_stats}
\alias{rd_stats}
\title{Stats from '.Rd' files}
\usage{
rd_stats(path)
}
\arguments{
\item{path}{Path to locally extracted tarball of package, obtained from
\link{extract_tarball}.}
}
\description{
Stats from '.Rd' files
}
