% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-stats.R
\name{rd_stats}
\alias{rd_stats}
\title{Stats from '.Rd' files}
\usage{
rd_stats(path)
}
\arguments{
\item{path}{Directory to source code of package being analysed}
}
\value{
A \code{data.frame} of function names and numbers of parameters and lines
of documentation for each, along with mean and median numbers of characters
used to document each parameter.
}
\description{
Stats from '.Rd' files
}
\examples{
f <- system.file ("extdata", "pkgstats_9.9.tar.gz", package = "pkgstats")
# have to extract tarball to call function on source code:
path <- extract_tarball (f)
rd_stats (path)
}
\seealso{
Other stats: 
\code{\link{desc_stats}()},
\code{\link{loc_stats}()},
\code{\link{pkgstats}()},
\code{\link{pkgstats_summary}()}
}
\concept{stats}
