% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctags.R
\name{tags_data}
\alias{tags_data}
\title{use ctags and gtags to parse call data}
\usage{
tags_data(path)
}
\arguments{
\item{path}{Path to local repository}
}
\description{
use ctags and gtags to parse call data
}
